# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs and applications
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: industry
          schema: { type: string }
        - in: query
          name: salaryMin
          schema: { type: integer }
        - in: query
          name: experienceLevel
          schema: { type: string }
        - in: query
          name: jobType
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [recent, salary, relevance] }
        - in: query
          name: savedOnly
          schema: { type: boolean }
      responses:
        "200":
          description: Jobs matching the search filters
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
    post:
      operationId: createJob
      tags: [jobs]
      summary: Add a job to the tracker
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404": { description: Job not found }
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a tracked job
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobUpdate" }
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Remove a job from the tracker
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List user skills
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Skill" }
    post:
      operationId: createSkill
      tags: [skills]
      summary: Add a skill
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SkillInput" }
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
  /skills/{id}:
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user portfolio projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      tags: [projects]
      summary: Add a project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, company, location, stage, nextAction, nextActionDate, postedAt, appliedAt, daysInStage, salary, experienceLevel, jobType, industry, saved, companyColor]
      properties:
        id: { type: integer }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        stage: { type: string, enum: [wishlist, applied, interview, offer, rejected] }
        nextAction: { type: string }
        nextActionDate: { type: string }
        postedAt: { type: string }
        appliedAt: { type: ["string", "null"] }
        daysInStage: { type: integer }
        salary: { type: string }
        experienceLevel: { type: string }
        jobType: { type: string }
        industry: { type: string }
        saved: { type: boolean }
        companyColor: { type: string }
        companyInitials: { type: string }
        description: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        contact: { type: ["string", "null"] }
        recruiterName: { type: ["string", "null"] }
        recruiterEmail: { type: ["string", "null"] }
        interviewDate: { type: ["string", "null"] }
        followUpDate: { type: ["string", "null"] }
        resumeVersion: { type: ["string", "null"] }
        skillsRequired:
          type: array
          items: { type: string }
    JobInput:
      type: object
      required: [title, company, location, stage]
      properties:
        title: { type: string, minLength: 1 }
        company: { type: string, minLength: 1 }
        location: { type: string }
        stage: { type: string, enum: [wishlist, applied, interview, offer, rejected] }
        nextAction: { type: string }
        nextActionDate: { type: string }
        salary: { type: string }
        experienceLevel: { type: string }
        jobType: { type: string }
        industry: { type: string }
        saved: { type: boolean }
        description: { type: string }
        url: { type: string }
        notes: { type: string }
        contact: { type: string }
        recruiterName: { type: string }
        recruiterEmail: { type: string }
        interviewDate: { type: string }
        followUpDate: { type: string }
        resumeVersion: { type: string }
        skillsRequired:
          type: array
          items: { type: string }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        company: { type: string }
        location: { type: string }
        stage: { type: string, enum: [wishlist, applied, interview, offer, rejected] }
        nextAction: { type: string }
        nextActionDate: { type: string }
        salary: { type: string }
        experienceLevel: { type: string }
        jobType: { type: string }
        industry: { type: string }
        saved: { type: boolean }
        description: { type: string }
        url: { type: string }
        notes: { type: string }
        contact: { type: string }
        recruiterName: { type: string }
        recruiterEmail: { type: string }
        interviewDate: { type: string }
        followUpDate: { type: string }
        resumeVersion: { type: string }
        skillsRequired:
          type: array
          items: { type: string }
...
    Skill:
      type: object
      required: [id, name, proficiency, yearsExperience]
      properties:
        id: { type: integer }
        name: { type: string }
        proficiency: { type: string, enum: [beginner, intermediate, advanced, expert] }
        yearsExperience: { type: integer }
    SkillInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        proficiency: { type: string, enum: [beginner, intermediate, advanced, expert] }
        yearsExperience: { type: integer }
    Project:
      type: object
      required: [id, name, skills, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        skills:
          type: array
          items: { type: string }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        createdAt: { type: string }
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        url: { type: string }
        skills:
          type: array
          items: { type: string }
        startDate: { type: string }
        endDate: { type: string }
...