# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillProficiency } from './skillProficiency';

export interface Skill {
  id: number;
  name: string;
  proficiency: SkillProficiency;
  yearsExperience: number;
}